updated types and data folders to follow correct conventions
