+++ src/data/StaffNavList.ts
@@ -1,12 +1,4 @@
-export type Staff = {
-  id: number;
-  name: string;
-  start: string;
-  courseEnd: string;
-  manager: string;
-  role: string;
-  description: string;
-};
+
 
 export const StaffList = [
   {
