+++ src/types/StaffNavTypes.ts
@@ -1,0 +1,9 @@
+export type Staff = {
+    id: number;
+    name: string;
+    start: string;
+    courseEnd: string;
+    manager: string;
+    role: string;
+    description: string;
+  };
